move types file to utils folder and update imports
