+++ frontend/src/app/features/userSlice.ts
@@ -1,5 +1,5 @@
 import { createSlice, PayloadAction } from "@reduxjs/toolkit";
-import { User } from './types'
+import { User } from '../../util/types'
 
 const initialState: User = {
   socketId: '',
