+++ frontend/src/util/middleware/socketActions/activeUsers.tsx
@@ -3,7 +3,7 @@
  */
 import { store } from "../../../app/store";
 import { getAllActiveUsers } from "../../../app/features/activeUsersSlice";
-import { User } from "../../../app/features/types";
+import { User } from "../../types";
 
 export const setActiveUsers = (users: User[]) => {
   store.dispatch(getAllActiveUsers(users))
